refactor: rename workflow approval files
